Separate Texturable from Visible Trait
