+++ src/main/groovy/com/scorsi/gameengine/entities/Texturable.groovy
@@ -1,0 +1,23 @@
+package com.scorsi.gameengine.entities
+
+import groovy.transform.SelfType
+
+import java.awt.Graphics
+import java.awt.image.BufferedImage
+
+@SelfType(Entity)
+trait Texturable implements Visible {
+
+    BufferedImage image
+
+    void render(Graphics g) {
+        if (clipToCamera)
+            g.drawImage(image,
+                    (position.x - handler.camera.offset.x) as Integer,
+                    (position.y - handler.camera.offset.y) as Integer,
+                    width as Integer, height as Integer, null)
+        else
+            g.drawImage(image, position.x as Integer, position.y as Integer, width as Integer, height as Integer, null)
+    }
+
+}
